refactor: Components, Layers 폴더 구조화 [#98]
- SiedBar 를 Layers 폴더로 이동
